fix: strip links from in app notifs
